Mutualize updateTrack method with reader and notification
